-refact

+++ src/SVExportDialog/src/export_dialog_impl.cpp
@@ -224,7 +224,8 @@ void ExportDialog::exportToXLSX(QString fileName) {
     xlsx.write(1, col, sd->name.c_str()); ++col;
 
     row = 2;
-    auto bsz = sd->buffData.size(); int cp = sd->buffBeginPos;
+    auto bsz = sd->buffData.size(); 
+    size_t cp = sd->buffBeginPos;
     while (cp != sd->buffValuePos) {
 
       uint64_t dt = sd->buffData[cp].beginTime;
@@ -311,7 +312,8 @@ void ExportDialog::exportToTXT(QString fileName) {
       out << "module = " << sd->module.c_str() << '\n';
       out << "signal = " << sd->name.c_str() << '\n';
 
-      auto bsz = sd->buffData.size(); int cp = sd->buffBeginPos;
+      auto bsz = sd->buffData.size(); 
+      size_t cp = sd->buffBeginPos;
       QString Value, time;
       while (cp != sd->buffValuePos) {
 
@@ -410,7 +412,8 @@ void ExportDialog::exportToJSON(QString fileName) {
     writer.Key("Signal");
     writer.String(sd->name.c_str());
 
-    auto bsz = sd->buffData.size(); int cp = sd->buffBeginPos;
+    auto bsz = sd->buffData.size();
+    size_t cp = sd->buffBeginPos;
     QString time, Value;
     while (cp != sd->buffValuePos) {
